+++ ad_map_access/impl/include/ad/map/route/Types.hpp
@@ -11,26 +11,9 @@
 #include "ad/map/point/Types.hpp"
 #include "ad/map/route/ConnectingRoute.hpp"
 #include "ad/map/route/FullRoute.hpp"
+#include "ad/map/route/FullRouteList.hpp"
 #include "ad/map/route/LaneChangeDirection.hpp"
 #include "ad/map/route/LaneInterval.hpp"
 #include "ad/map/route/LaneSegment.hpp"
 #include "ad/map/route/RoadSegment.hpp"
 #include "ad/map/route/RouteParaPoint.hpp"
-
-namespace ad {
-/*!
- * @brief namespace map
- */
-namespace map {
-/*!
- * @brief namespace lane
- *
- * Handling of lanes
- */
-namespace route {
-
-typedef std::vector<ad::map::route::FullRoute> FullRouteList;
-
-} // namespace lane
-} // namespace map
-} // namespace ad