+++ doc/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Latest changes
 
-## Release 2.3.1
+## Release 2.4.0
 #### :rocket: New Features
 * Qgis 2.18 to QGis 3.14
    - The plugin code is adapted to QGis3.14, Python 3 and the python bindings of the new C++ interface.
@@ -9,8 +9,19 @@
    - Settings for Route Prediction test,Routing Test at run time.
    - For open drive map, the attributes roadId, laneSectionIndex and laneIndex are added.
    - The ENU coordinates of the map snapped point is added as an attribute.
+* Added AdMapMatching::findRouteLanes()
+* Added lane::findNearestPointOnLaneInterval()
+* Added matchRadius and samplingDistance members to match::MapMatchedObjectBoundingBox
+
 #### :ghost: Maintenance
 * Use target python version for build
+* Generate list types and tests: ENUBorderList, GeoBorderList, ECEFBorderList,
+  FullRouteList, MapMatchedObjectReferencePositionList
+* Fixed issue in reading of some OpenDRIVE maps
+* Fixed point::getParametricRange() for degenerated edges
+* Fixed lane::calcLength(<ENU,ECEF,Geo>BorderList) functions
+* Allow rounding errors in lane::isPointWithinBorderPoints() (might e.g. happen after projection of outside points onto the borders)
+* Fix route::planning::createRoutingPoint(match::LaneOccupiedRegion)
 
 ## Release 2.3.0
 #### :rocket: New Features